Update autocomplete demos for SB
