--- libs/documentation/src/lib/components/autocomplete/demos/aria-label/autocomplete-aria-label.module.ts
@@ -1,13 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { SdsAutocompleteModule } from '@gsa-sam/components';
-import { AutocompleteAriaLabelComponent } from './autocomplete-aria-label.component';
-import { FormsModule } from '@angular/forms';
-
-@NgModule({
-  imports: [CommonModule, FormsModule, SdsAutocompleteModule],
-  declarations: [AutocompleteAriaLabelComponent],
-  exports: [AutocompleteAriaLabelComponent],
-  bootstrap: [AutocompleteAriaLabelComponent],
-})
-export class AutocompleteAriaLabelModule {}
