+++ libs/documentation/src/lib/components/autocomplete/demos/autocomplete-free-text/autocomplete-free-text.component.html
@@ -1,0 +1,10 @@
+<div class="margin-bottom-4">
+    <p>tagMode</p>
+    <sds-autocomplete [service]="service" [(ngModel)]="model" [configuration]="settings"> </sds-autocomplete>
+</div>
+{{ model | json }}
+<div class="margin-bottom-4">
+    <p>free text</p>
+    <sds-autocomplete [service]="service" [(ngModel)]="model2" [configuration]="settings2"> </sds-autocomplete>
+</div>
+{{ model2 | json }}
