+++ libs/documentation/src/lib/components/autocomplete/demos/autocomplete-grouping/autocomplete-grouping.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SdsAutocompleteModule } from '@gsa-sam/components';
+import { FormsModule } from '@angular/forms';
+import { AutocompleteGrouping } from './autocomplete-grouping.component';
+
+@NgModule({
+  declarations: [AutocompleteGrouping],
+  imports: [CommonModule, FormsModule, SdsAutocompleteModule],
+  exports: [AutocompleteGrouping],
+  bootstrap: [AutocompleteGrouping],
+})
+export class AutocompleteGroupingModule {}
