+++ libs/documentation/src/lib/components/autocomplete/demos/autocomplete-selection-mode/autocomplete-selection-mode.component.html
@@ -1,0 +1,24 @@
+<div class="margin-bottom-4">
+    <p>Single Selection Mode</p>
+    <sds-autocomplete
+      [service]="service"
+      (ngModelChange)="changes($event)"
+      [(ngModel)]="model"
+      [configuration]="settings"
+    >
+    </sds-autocomplete>
+  </div>
+  {{ model | json }}
+  
+  <div class="margin-bottom-4">
+    <p>Multiple Selection Mode</p>
+    <sds-autocomplete
+      [service]="service"
+      (ngModelChange)="changes($event)"
+      [(ngModel)]="model3"
+      [configuration]="settings3"
+    ></sds-autocomplete>
+  </div>
+  
+  {{ model3 | json }}
+
