+++ libs/documentation/src/lib/components/autocomplete/demos/autocomplete-selection-mode/autocomplete-selection-mode.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SdsAutocompleteModule } from '@gsa-sam/components';
+import { FormsModule } from '@angular/forms';
+import { AutocompleteSelectionMode } from './autocomplete-selection-mode.component';
+
+@NgModule({
+  declarations: [AutocompleteSelectionMode],
+  imports: [CommonModule, FormsModule, SdsAutocompleteModule],
+  exports: [AutocompleteSelectionMode],
+  bootstrap: [AutocompleteSelectionMode],
+})
+export class AutocompleteSelectionModeModule {}
