--- libs/documentation/src/lib/components/autocomplete/demos/basic/autocomplete-basic.component.html
@@ -1,31 +1,0 @@
-<div class="margin-bottom-4">
-  <h4 id="autoBasicLabel" for="autocomplete1">Basic Autocomplete with Single Selection mode</h4>
-  <p>
-    <code>this.settings.selectionMode = SelectionMode.SINGLE</code> is used to set the selection mode of the
-    autocomplete.
-  </p>
-  <sds-autocomplete
-    [service]="service"
-    (ngModelChange)="changes($event)"
-    [(ngModel)]="model"
-    [configuration]="settings"
-  >
-  </sds-autocomplete>
-</div>
-{{ model | json }}
-
-<div class="margin-bottom-4">
-  <h4 id="autoBasic2Label" for="autocomplete1">Basic Autocomplete with Multi Selection mode</h4>
-  <p>
-    <code> this.settings3.selectionMode = SelectionMode.MULTIPLE</code> is used to set the selection mode of the
-    autocomplete.
-  </p>
-  <sds-autocomplete
-    [service]="service"
-    (ngModelChange)="changes($event)"
-    [(ngModel)]="model3"
-    [configuration]="settings3"
-  ></sds-autocomplete>
-</div>
-
-{{ model3 | json }}
