+++ libs/documentation/src/lib/components/autocomplete/demos/configurable/autocomplete-configurable.component.html
@@ -1,0 +1,11 @@
+<div class="margin-bottom-4">
+  <sds-autocomplete
+    [service]="service"
+    (ngModelChange)="changes($event)"
+    [(ngModel)]="model"
+    [configuration]="settings"
+    [disabled]="disabled"
+  >
+  </sds-autocomplete>
+</div>
+{{ model | json }}
