+++ libs/documentation/src/lib/components/autocomplete/demos/configurable/autocomplete-configurable.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AutocompleteConfigurable } from './autocomplete-configurable.component';
+import { SdsAutocompleteModule } from '@gsa-sam/components';
+import { FormsModule } from '@angular/forms';
+
+@NgModule({
+  declarations: [AutocompleteConfigurable],
+  imports: [CommonModule, FormsModule, SdsAutocompleteModule],
+  exports: [AutocompleteConfigurable],
+  bootstrap: [AutocompleteConfigurable],
+})
+export class AutocompleteConfigurableModule {}
