--- libs/documentation/src/lib/components/autocomplete/demos/disable/autocomplete-disable.component.html
@@ -1,23 +1,0 @@
-<div>
-  <h4 id="autoOptional6Label">Autocomplete with disable with single selection mode</h4>
-  <p><code>[disabled]="true" </code> is used to set the disable the control.</p>
-  <sds-autocomplete
-    [service]="service"
-    [disabled]="true"
-    [(ngModel)]="multiModeDisableModel"
-    [configuration]="multiModeDisableSettings"
-  >
-  </sds-autocomplete>
-</div>
-
-<div>
-  <h4 id="autoOptionalDisableLabel">Autocomplete with disable with multiple selection mode</h4>
-  <p><code>[disabled]="true" </code> is used to set the disable the control.</p>
-  <sds-autocomplete
-    [service]="service"
-    [disabled]="true"
-    [(ngModel)]="disableMModel"
-    [configuration]="disableMSettings"
-  >
-  </sds-autocomplete>
-</div>
