--- libs/documentation/src/lib/components/autocomplete/demos/group/autocomplete-group.component.html
@@ -1,7 +1,0 @@
-<div class="margin-bottom-4">
-  <h4 id="autoGroupLabel" for="autocomplete1">Autocomplete with Group</h4>
-  <p><code>this.settings.isGroupingEnabled = true</code> is used to set the enable the grouping.</p>
-  <p><code>this.settings.groupByChild = 'elements';</code> child property name for grouping.</p>
-  <sds-autocomplete [service]="service" [(ngModel)]="model" [configuration]="settings"> </sds-autocomplete>
-</div>
-{{ model | json }}
