--- libs/documentation/src/lib/components/autocomplete/demos/group/autocomplete-group.component.ts
@@ -1,28 +1,0 @@
-import { Component } from '@angular/core';
-import { SDSAutocompletelConfiguration, SDSSelectedItemModel, SelectionMode } from '@gsa-sam/components';
-import { AutocompleteSampleDataService } from './autocomplete-sample.service';
-
-@Component({
-  templateUrl: './autocomplete-group.component.html',
-  selector: `sds-autocomplete-group-demo`,
-  providers: [AutocompleteSampleDataService],
-})
-export class AutocompleteGroup {
-  public settings = new SDSAutocompletelConfiguration();
-  public model = new SDSSelectedItemModel();
-
-  constructor(public service: AutocompleteSampleDataService) {
-    this.setup();
-  }
-
-  setup() {
-    this.settings.id = 'autocompleteGroup';
-    this.settings.primaryKeyField = 'element_id';
-    this.settings.primaryTextField = 'value';
-    this.settings.labelText = 'Autocomplete 1';
-    this.settings.selectionMode = SelectionMode.MULTIPLE;
-    this.settings.autocompletePlaceHolderText = 'eg: Formula Grants';
-    this.settings.isGroupingEnabled = true;
-    this.settings.groupByChild = 'elements';
-  }
-}
