--- libs/documentation/src/lib/components/autocomplete/demos/inputreadonly/autocomplete-inputreadonly.component.html
@@ -1,10 +1,0 @@
-<div class="margin-bottom-4">
-  <h4 id="autoOptional2Label">Autocomplete with dropdown input read only</h4>
-  <p><code> this.inputReadOnlySettings.inputReadOnly = true</code> to enable input readonly and work as select</p>
-  <sds-autocomplete
-    [service]="service"
-    (ngModelChange)="changes($event)"
-    [(ngModel)]="inputReadOnlyModel"
-    [configuration]="inputReadOnlySettings"
-  ></sds-autocomplete>
-</div>
