--- libs/documentation/src/lib/components/autocomplete/demos/mincharacter/autocomplete-mincharacter.component.html
@@ -1,9 +1,0 @@
-<div class="margin-bottom-4">
-  <h4 id="autoOptional5Label">Autocomplete with minimum character count</h4>
-  <p>
-    <code>this.mincharacterSettings.minimumCharacterCountSearch = 3</code> is used to enable the minimum character
-    count.
-  </p>
-  <sds-autocomplete [service]="service" [(ngModel)]="mincharacterModel" [configuration]="mincharacterSettings">
-  </sds-autocomplete>
-</div>
