--- libs/documentation/src/lib/components/autocomplete/demos/selectgroup/autocomplete-selectgroup.component.html
@@ -1,9 +1,0 @@
-<div class="margin-bottom-4">
-  <h4 id="autoSelectGroupLabel" for="autocompleteSelectGroup">Autocomplete with Group as label</h4>
-  <p><code>this.settings.isGroupingEnabled = true</code> is used to set the enable the grouping.</p>
-  <p><code>this.settings.groupByChild = 'elements';</code> child property name for grouping.</p>
-
-  <p><code>this.settings.isSelectableGroup = false;</code> to disable selecting group element</p>
-  <sds-autocomplete [service]="service" [(ngModel)]="model" [configuration]="settings"> </sds-autocomplete>
-</div>
-{{ model | json }}
