--- libs/documentation/src/lib/components/autocomplete/demos/selectgroup/data.ts
@@ -1,467 +1,0 @@
-export let GroupAutocompleteData = [
-  {
-    code: 'A',
-    elements: [
-      {
-        code: 'A',
-        elements: null,
-        description: null,
-        element_id: '0001001',
-        value: 'Formula Grants',
-      },
-      {
-        code: 'A',
-        elements: null,
-        description: null,
-        element_id: '0001002',
-        value: 'Formula Grants (Apportionments)',
-      },
-      {
-        code: 'A',
-        elements: null,
-        description: null,
-        element_id: '0001003',
-        value: 'Formula Grants (Cooperative Agreements)',
-      },
-      {
-        code: 'A',
-        elements: null,
-        description: null,
-        element_id: '0001004',
-        value: 'Formula Grants (Health Incentive Grants)',
-      },
-    ],
-    description:
-      'Formula Grants - Allocations of money to States or their subdivisions in accordance with a distribution formula prescribed by law or administrative regulation, for activities of a continuing nature not confined to a specific project.',
-    element_id: '0001',
-    value: 'FORMULA GRANTS',
-  },
-  {
-    code: 'B',
-    elements: [
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003001',
-        value: 'Cooperative Agreements',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003002',
-        value: 'Cooperative Agreements (Discretionary Grants)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003003',
-        value: 'Project Grants',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003004',
-        value: 'Project Grants (Capacity Building and Complaint Processing, Training)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003005',
-        value: 'Project Grants (Contracts)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003006',
-        value: 'Project Grants (Cooperative Agreements or Contracts)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003007',
-        value: 'Project Grants (Cooperative Agreements)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003008',
-        value: 'Project Grants (Discretionary)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003009',
-        value: 'Project Grants (Fellowships)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003010',
-        value: 'Project Grants (for administration projects authorized under Section',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003011',
-        value: 'Project Grants (for collaborative design of curriculum, production of',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003012',
-        value: 'Project Grants (for specified projects)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003013',
-        value: 'Project Grants (including individual awards)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003014',
-        value: 'Project Grants (including travel grants)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003015',
-        value: 'Project Grants (Rehabilitation of existing indoor and outdoor recreat',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003016',
-        value: 'Project Grants (Special)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003017',
-        value: 'Project Grants (to capitalize loan funds)',
-      },
-      {
-        code: 'B',
-        elements: null,
-        description: null,
-        element_id: '0003018',
-        value: 'Project Grants (with Formula Distribution)',
-      },
-    ],
-    description:
-      'Project Grants - The funding, for fixed or known periods, of specific projects or the delivery of specific services or products without liability for damages for failure to perform. Project grants include fellowships, scholarships, research grants, training grants, traineeships, experimental and demonstration grants, evaluation grants, planning grants, technical assistance grants, survey grants, construction grants, and unsolicited contractual agreements.',
-    element_id: '0003',
-    value: 'PROJECT GRANTS',
-  },
-  {
-    code: 'C',
-    elements: [
-      {
-        code: 'C',
-        elements: null,
-        description: null,
-        element_id: '0005001',
-        value: 'Direct Payments for Specified Use',
-      },
-      {
-        code: 'C',
-        elements: null,
-        description: null,
-        element_id: '0005002',
-        value: 'Direct Payments for Specified Use (Cooperative Agreements)',
-      },
-    ],
-    description:
-      'Direct Payments for a Specified Use - Financial assistance provided directly to individuals, private firms, and other private institutions to encourage or subsidize a particular activity by conditioning the receipt of the on a particular performance by the recipient. This does not include solicited contracts for the procurement of goods and services for the Federal government.',
-    element_id: '0005',
-    value: 'DIRECT PAYMENTS FOR A SPECIFIED USE',
-  },
-  {
-    code: 'D',
-    elements: [
-      {
-        code: 'D',
-        elements: null,
-        description: null,
-        element_id: '0007001',
-        value: 'Direct Payments with Unrestricted Use',
-      },
-    ],
-    description:
-      'Direct Payments with Unrestricted Use - Financial assistance provided directly to beneficiaries who satisfy eligibility requirements with no restrictions imposed on the recipient as to how the money is spent. Included are payments under retirement, pension, and compensation programs.',
-    element_id: '0007',
-    value: 'DIRECT PAYMENTS WITH UNRESTRICTED USE',
-  },
-  {
-    code: 'E',
-    elements: [
-      {
-        code: 'E',
-        elements: null,
-        description: null,
-        element_id: '0009001',
-        value: 'Direct Loans',
-      },
-    ],
-    description:
-      'Direct Loans - Financial assistance provided through the lending of Federal monies for a specific period of time, with a reasonable expectation of repayment. Such loans may or may not require the payment of interest.',
-    element_id: '0009',
-    value: 'DIRECT LOANS',
-  },
-  {
-    code: 'F',
-    elements: [
-      {
-        code: 'F',
-        elements: null,
-        description: null,
-        element_id: '0011001',
-        value: 'Guaranteed/Insured Loans',
-      },
-      {
-        code: 'F',
-        elements: null,
-        description: null,
-        element_id: '0011002',
-        value: 'Guaranteed/Insured Loans (including Guaranty Participation Loans)',
-      },
-      {
-        code: 'F',
-        elements: null,
-        description: null,
-        element_id: '0011003',
-        value: 'Guaranteed/Insured Loans (including Immediate Participation Loans)',
-      },
-      {
-        code: 'F',
-        elements: null,
-        description: null,
-        element_id: '0011004',
-        value: 'Loan Guarantees/Grants',
-      },
-    ],
-    description:
-      'Guaranteed/Insured Loans - Programs in which the Federal Government makes an arrangement to indemnify a lender against all or part of all of any defaults by those responsible for the repayment of loans.',
-    element_id: '0011',
-    value: 'GUARANTEED/INSURED LOANS',
-  },
-  {
-    code: 'G',
-    elements: [
-      {
-        code: 'G',
-        elements: null,
-        description: null,
-        element_id: '0013001',
-        value: 'Insurance',
-      },
-      {
-        code: 'G',
-        elements: null,
-        description: null,
-        element_id: '0013002',
-        value: 'Insurance (Guaranteed Surety Bonds)',
-      },
-      {
-        code: 'G',
-        elements: null,
-        description: null,
-        element_id: '0013003',
-        value: 'Insurance (Reimbursement)',
-      },
-    ],
-    description:
-      'Insurance - Financial assistance provided to assure reimbursement for losses sustained under specified conditions. Coverage may be provided directly by the Federal Government or through private carriers and may or may not involve the payment of premiums.',
-    element_id: '0013',
-    value: 'INSURANCE',
-  },
-  {
-    code: 'H',
-    elements: [
-      {
-        code: 'H',
-        elements: null,
-        description: null,
-        element_id: '0015001',
-        value: 'Sale, Exchange, or Donation of Property and Goods',
-      },
-    ],
-    description:
-      'Sale, Exchange, or Donation of Property or Goods - Programs that provide for the sale, exchange, or donation of Federal real property, personal property, commodities, and other goods including land, buildings, equipment, food, and drugs. This does not include the loan of, use of, or access to Federal facilities or property.',
-    element_id: '0015',
-    value: 'SALE, EXCHANGE, OR DONATION OF PROPERTY OR GOODS',
-  },
-  {
-    code: 'I',
-    elements: [
-      {
-        code: 'I',
-        elements: null,
-        description: null,
-        element_id: '0017001',
-        value: 'Use of Property, Facilities, and Equipment',
-      },
-    ],
-    description:
-      'Use of Property, Facilities, or Equipment - Programs that provide for the loan of, use of, or access to Federal facilities or property wherein the federally owned facilities or property do not remain in the possession of the recipient of the assistance.',
-    element_id: '0017',
-    value: 'USE OF PROPERTY, FACILITIES, OR EQUIPMENT',
-  },
-  {
-    code: 'J',
-    elements: [
-      {
-        code: 'J',
-        elements: null,
-        description: null,
-        element_id: '0019001',
-        value: 'Provision of Specialized Services',
-      },
-    ],
-    description:
-      'Provision of Specialized Services - Programs that provide Federal personnel to directly perform certain tasks for the benefit of communities or individuals. These services may be performed in conjunction with non-Federal personnel, but they involve more than consultation, advice, or counseling.',
-    element_id: '0019',
-    value: 'PROVISION OF SPECIALIZED SERVICES',
-  },
-  {
-    code: 'K',
-    elements: [
-      {
-        code: 'K',
-        elements: null,
-        description: null,
-        element_id: '0021001',
-        value: 'Advisory Services and Counseling',
-      },
-      {
-        code: 'K',
-        elements: null,
-        description: null,
-        element_id: '0021002',
-        value: 'Advisory Services on Compliance',
-      },
-    ],
-    description:
-      'Advisory Services and Counseling - Programs that provide Federal specialists to consult, advise, or counsel communities or individuals including conferences, workshops, or personal contacts. This may involve the use of published information, but only in a secondary capacity.',
-    element_id: '0021',
-    value: 'ADVISORY SERVICES AND COUNSELING',
-  },
-  {
-    code: 'L',
-    elements: [
-      {
-        code: 'L',
-        elements: null,
-        description: null,
-        element_id: '0023001',
-        value: 'Dissemination of Technical Information',
-      },
-      {
-        code: 'L',
-        elements: null,
-        description: null,
-        element_id: '0023002',
-        value: 'Information',
-      },
-    ],
-    description:
-      'Dissemination of Technical Information - Programs that provide for the publication and distribution of information or data of a specialized technical nature, frequently through clearinghouses or libraries. This does not include conventional public information services designed for general public consumption.',
-    element_id: '0023',
-    value: 'DISSEMINATION OF TECHNICAL INFORMATION',
-  },
-  {
-    code: 'M',
-    elements: [
-      {
-        code: 'M',
-        elements: null,
-        description: null,
-        element_id: '0025001',
-        value: 'Training',
-      },
-    ],
-    description:
-      'Training - Programs that provide instructional activities conducted directly by a Federal agency for individuals not employed by the Federal Government.',
-    element_id: '0025',
-    value: 'TRAINING',
-  },
-  {
-    code: 'N',
-    elements: [
-      {
-        code: 'N',
-        elements: null,
-        description: null,
-        element_id: '0027020',
-        value: 'Resolution of Disputes',
-      },
-      {
-        code: 'N',
-        elements: null,
-        description: null,
-        element_id: '0027001',
-        value: 'Investigation of Complaints',
-      },
-      {
-        code: 'N',
-        elements: null,
-        description: null,
-        element_id: '0027002',
-        value: 'Investigation of Complaints (Compliance Reviews)',
-      },
-    ],
-    description:
-      'Investigation of Complaints - Federal administrative agency activities that are initiated in response to requests, either formal or informal, to examine or investigate claims of violations of Federal statutes, policy, or procedure. Such claims must come from outside the Federal Government.',
-    element_id: '0027',
-    value: 'INVESTIGATION OF COMPLAINTS',
-  },
-  {
-    code: 'O',
-    elements: [
-      {
-        code: 'O',
-        elements: null,
-        description: null,
-        element_id: '0029001',
-        value: 'Federal Employment',
-      },
-    ],
-    description:
-      'Federal Employment - Programs that reflect the government-wide resposibilities of the Office of Personnel Management in the recruitment and hiring of Federal civilian agency personnel.',
-    element_id: '0029',
-    value: 'FEDERAL EMPLOYMENT',
-  },
-  {
-    code: 'Z',
-    elements: [
-      {
-        code: 'Z',
-        elements: null,
-        description: null,
-        element_id: '0090001',
-        value: 'Salaries and Expenses',
-      },
-    ],
-    description: null,
-    element_id: '0090',
-    value: 'SALARIES AND EXPENSES',
-  },
-];
